refactor: survey intro 구조 변경
