+++ src/features/review/steps/Intro.tsx
@@ -14,8 +14,8 @@ import useDidMount from '~/hooks/lifeCycle/useDidMount';
 import useStep from '~/hooks/step/useStep';
 import recordEvent from '~/utils/event';
 
+import IntroHeader from '../../../components/intro/IntroHeader';
 import { fixedBottomCss } from '../style';
-import IntroHeader from './intro/IntroHeader';
 import { type StepProps } from './type';
 
 interface Props extends StepProps {
